Add authorization middleware

+++ routes/articles.js
@@ -6,6 +6,7 @@ var User = require("../models/User");
 var Comment = require("../models/Comment");
 var jwt = require("../modules/token");
 var { profileInfo } = require("./profiles");
+var { verifyAuthor } = require("../midlewares/auth");
 
 // get all articles
 router.get("/", jwt.verifyToken, async (req, res, next) => {
@@ -72,7 +73,7 @@ router.get("/feed", jwt.verifyToken, async (req, res, next) => {
 
 // create article
 router.post("/", jwt.verifyToken, async (req, res, next) => {
-    var slug = req.body.article.title.toLowerCase().replaceAll(" ", "-");
+    var slug = req.body.article.title.trim().toLowerCase().replaceAll(" ", "-");
     req.body.article.slug = slug;
     req.body.article.author = req.user.id;
 
@@ -100,14 +101,16 @@ router.get("/:slug", async (req, res, next) => {
 
 // update article
 router.put("/:slug", jwt.verifyToken, async (req, res, next) => {
-    if (req.body.article.title) {
-        var slug = req.body.article.title.toLowerCase().replaceAll(" ", "-");
-        req.body.article.slug = slug;
-    }
-
-    var slug = req.params.slug;
-
     try {
+        var slug = req.params.slug;
+        var currentArticle = await Article.findOne({ slug });
+        verifyAuthor(currentArticle.author, req.user.id, res);
+
+        if (req.body.article.title) {
+            let slug = req.body.article.title.toLowerCase().replaceAll(" ", "-");
+            req.body.article.slug = slug;
+        }
+
         var article = await Article.findOneAndUpdate({ slug }, req.body.article, { new : true })
             .populate("author");
 
@@ -121,6 +124,8 @@ router.put("/:slug", jwt.verifyToken, async (req, res, next) => {
 router.delete("/:slug", jwt.verifyToken, async (req, res, next) => {
     var slug = req.params.slug;
     try {
+        var currentArticle = await Article.findOne({ slug });
+        verifyAuthor(currentArticle.author, req.user.id, res);
         var article = await Article.findOneAndDelete({ slug });
         res.status(200).json({ message : "Article deleted successfully" });
     } catch (error) {
@@ -134,8 +139,8 @@ router.get("/:slug/comments", jwt.verifyToken, async (req, res, next) => {
     var slug = req.params.slug;
 
     try {
-        var article = await Article.findOne({ slug }).populate("comments");
-        res.status(200).json({ comments : article.comments });
+        var article = await Article.findOne({ slug }).populate({ path : "comments", populate : "author"});
+        res.status(200).json({ comments : article.comments.map(comment => commentInfo(comment, req.user)) });
     } catch (error) {
         next(error);
     }
@@ -148,22 +153,24 @@ router.post("/:slug/comments", jwt.verifyToken, async (req, res, next) => {
     req.body.comment.author = author;
 
     try {
-        var comment = await Comment.create(req.body.comment);
+        var comment = await (await Comment.create(req.body.comment)).execPopulate("author");
         var article = await Article.findOneAndUpdate(
             { slug }, { $addToSet : { comments : comment.id } });
 
-        res.status(200).json({ comment });
+        res.status(200).json({ comment : commentInfo(comment, req.user) });
     } catch (error) {
         next(error);
     }
 });
 
-// create a comment
+// dalete a comment
 router.delete("/:slug/comments/:id", jwt.verifyToken, async (req, res, next) => {
     var slug = req.params.slug;
     var commentId = req.params.id;
 
     try {
+        var currentComment = await Comment.findById(commentId);
+        verifyAuthor(currentComment.author, req.user.id, res);
         var comment = await Comment.findByIdAndDelete(commentId);
         var article = await Article.findOneAndUpdate(
             { slug }, { $pull : { comments : commentId } });
@@ -218,6 +225,16 @@ function articleInfo(article, currentUser) {
         favoritesCount : article.favorites.length,
         createdAt : article.createdAt,
         updatedAt : article.updatedAt
+    }
+}
+
+function commentInfo(comment, currentUser) {
+    return {
+        id : comment.id,
+        createdAt : comment.createdAt,
+        updatedAt : comment.updatedAt,
+        body : comment.body,
+        author : profileInfo(comment.author, currentUser)
     }
 }
